# OpenApiSpec.yml
openapi: 3.0.0
servers:
  - url: 'https://localhost:8888/api/v1/'
info:
  description: >-
    This is the public JSON REST API for the Legal API.
  version: 1.0.0
  title: Legal API 
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /api/v1/upload:
    post:
      tags:
        - Documents
      summary: Upload a Document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to upload
                groupID:
                  type: string
                  description: Optional group ID
                name:
                  type: string
                  description: Optional document name
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/documents:
    get:
      tags:
        - Documents
      summary: Get all documents for user's group
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
  /api/v1/documents/{documentID}:
    get:
      tags:
        - Documents
      summary: Get a specific document
      security:
        - bearerAuth: []
      parameters:
        - name: documentID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Documents
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - name: documentID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '403':
          description: Access forbidden - Requires Admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: object
            properties:
              location:
                type: string
              param:
                type: string
              msg:
                type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupID:
          type: string
        name:
          type: string
        filePath:
          type: string
        size:
          type: number
        mimeType:
          type: string